Fixing bug for monthly_day dates for December
Next recurrences falling in December would be accidentally skipped to
the end of the month due to the monthly rollover behaviour.

i.e. a monthly repeating event for the 4th of every month for the 4th of
November would evalutate the next date incorrectly as December 31st
instead of December 4th.